FIX: Images Folder, FEAT: Toggle SP Status
All Images will be stored in the Images folder in their respective directories

+++ routes/serviceProviderProfile.js
@@ -1,3 +1,4 @@
+
 import express  from "express";
 import ServiceProviders from "../modals/ServiceProviders.js";
 import multer from "multer";
@@ -9,17 +10,36 @@ import path from 'path';
 import {fileURLToPath} from 'url';
 import Cryptr from "cryptr"
 import Review from "../modals/Review.js";
+import Jobs from "../modals/Jobs.js";
 const router = express.Router();
 
-const storage = multer.diskStorage({
+const storageProfilePicture = multer.diskStorage({
     destination: function(req, file, cb) {
-      cb(null, './uploads/');
+      cb(null, './Images/ProfilePics');
     },
     filename: function(req, file, cb) {
-      cb(null, file.originalname)?.toLowerCase().split(' ').join('_');
+      var date = new Date();
+      var timeStamp = date.toDateString()+"_"+date.getHours()+"-"+date.getMinutes()+"-"+date.getSeconds()+"-"+date.getMilliseconds();
+      // timeStampArray.push(timeStamp)
+      timeStamp = timeStamp?.toLowerCase().split(' ').join('_');
+      file.originalname = file.originalname?.toLowerCase().split(' ').join('_');
+      cb(null , timeStamp+"_"+file.originalname)
     }
   });
 
+const storagePortfolioPicture = multer.diskStorage({
+    destination: function(req, file, cb) {
+      cb(null, './Images/PortfolioPics/');
+    },
+    filename: function(req, file, cb) {
+      var date = new Date();
+      var timeStamp = date.toDateString()+"_"+date.getHours()+"-"+date.getMinutes()+"-"+date.getSeconds()+"-"+date.getMilliseconds();
+      // timeStampArray.push(timeStamp)
+      timeStamp = timeStamp?.toLowerCase().split(' ').join('_');
+      file.originalname = file.originalname?.toLowerCase().split(' ').join('_');
+      cb(null , timeStamp+"_"+file.originalname)
+    }
+  });
   const fileFilter = (req, file, cb) => {
     if (file.mimetype === 'image/jpeg' || file.mimetype === 'image/png' || file.mimetype === 'image/jpg' || file.mimetype === 'image/jfif') {
       cb(null, true);
@@ -27,8 +47,15 @@ const storage = multer.diskStorage({
       cb(null, false);
     }
   };
-const uploads = multer({
-    storage: storage,
+const uploadPortfolioPicture = multer({
+    storage: storagePortfolioPicture,
+    limits: {
+      fileSize: 1024 * 1024 * 5
+    },
+    fileFilter: fileFilter
+  });
+  const uploadProfilePicture = multer({
+    storage: storageProfilePicture,
     limits: {
       fileSize: 1024 * 1024 * 5
     },
@@ -75,7 +102,7 @@ router.get("/:email" ,async function(req,res){
     }) 
 })
 
-router.put("/updateProfile" ,uploads.single("url"),(req, res)=>{
+router.put("/updateProfile" ,uploadProfilePicture.single("url"),(req, res)=>{
   let serviceId;
   let userId;
   console.log("Request received to update profile" , req.body)
@@ -105,10 +132,15 @@ router.put("/updateProfile" ,uploads.single("url"),(req, res)=>{
             console.log(3)
             console.log(updatedSP)
             const __filename = fileURLToPath(import.meta.url);
-            fs.unlink(path.dirname(path.dirname(__filename))+"/"+updatedSP.profilePicture , (err)=>{
-              if(err) console.log(err);
-              console.log('image was deleted');
-            })
+            // console.log(__filename)
+            // console.log(path.dirname(path.dirname(__filename)))
+            console.log(updatedSP.profilePicture)
+            if (updatedSP.profilePicture != "uploads\\defaultProfile.png"){
+              fs.unlink(path.dirname(path.dirname(__filename))+"/"+updatedSP.profilePicture , (err)=>{
+                if(err) console.log(err);
+                console.log('image was deleted');
+              })
+            }
            })
           }
           else{
@@ -127,8 +159,10 @@ router.put("/updateProfile" ,uploads.single("url"),(req, res)=>{
     })
 
   }).catch(err=>console.log(err))
-
-  res.status(200).send("");
+  setTimeout(() => {
+    res.status(200).send("");
+    
+  }, 1000);
   
 })
 router.put("/updateProfile/deleteImage" , (req , res)=>{
@@ -148,7 +182,7 @@ router.put("/updateProfile/deleteImage" , (req , res)=>{
     }).then((sp)=>res.status(200).send(sp)).catch((err)=>res.status(500).send(err))
   }).catch((err)=>res.status(500).send(err))
 })
-router.put("/updateProfile/uploadImage" ,uploads.single("url"),(req, res)=>{
+router.put("/updateProfile/uploadImage" ,uploadPortfolioPicture.single("url"),(req, res)=>{
   let email = req.body.email
   let imageUrl = req.file.path
   console.log(imageUrl)
@@ -197,6 +231,18 @@ router.get("/reviews/:id",async(req, res)=>{
       res.status(200).send(data)
     }
   })
+})
+router.patch("/sp/status-update",(req , res)=>{
+  var currentStatus = req.body.status
+  console.log( "REQUEST RECEIVED TO TOGGLE STATUS OF", req.body.id)
+  console.log(currentStatus)
+  ServiceProviders.findOneAndUpdate({serviceProvider:req.body?.id},
+    [{$set:{status:
+      {
+        $cond: { if:{$or:[{ $eq: [ currentStatus, "" ] } ,{ $eq: [ currentStatus, "Enabled" ] } ]} , then: "Disabled", else: "Enabled" }
+      }
+  }}]
+    ).then((sp)=>res.status(200).send(sp)).catch((err)=>{res.status(500).send(err) ; console.log(err)} )
 })
